refactor: organise CLI code
