+++ backend/pyspur/cli/__init__.py
@@ -1,0 +1,5 @@
+"""PySpur CLI package."""
+
+from .main import main
+
+__all__ = ["main"]
